feat(brokers): Remove support for context=fx-desktop-v1
fx-context-v2 has been supported for 3 years now, and fx-desktop-v1
is hanging around like a bad smell.

Users who attempt to load FxA with `context=fx_desktop_v1` are
redirected to an `/update_firefox` screen.

The fx_desktop_v1 channel related files have been renamed to fx_ios_v1

BREAKING CHANGE

fixes #5651
