+++ app/tests/test_start.js
@@ -16,7 +16,7 @@ require('./spec/lib/app-start');
 require('./spec/lib/assertion');
 require('./spec/lib/auth-errors');
 require('./spec/lib/channels/duplex');
-require('./spec/lib/channels/fx-desktop-v1');
+require('./spec/lib/channels/fx-ios-v1');
 require('./spec/lib/channels/iframe');
 require('./spec/lib/channels/inter-tab');
 require('./spec/lib/channels/notifier');
@@ -24,7 +24,7 @@ require('./spec/lib/channels/notifier-mixin');
 require('./spec/lib/channels/null');
 require('./spec/lib/channels/receivers/postmessage');
 require('./spec/lib/channels/receivers/web-channel');
-require('./spec/lib/channels/senders/fx-desktop-v1');
+require('./spec/lib/channels/senders/fx-ios-v1');
 require('./spec/lib/channels/senders/web-channel');
 require('./spec/lib/channels/web');
 require('./spec/lib/cocktail');
@@ -80,7 +80,6 @@ require('./spec/lib/xss');
 require('./spec/models/account');
 require('./spec/models/attached-clients');
 require('./spec/models/auth_brokers/base');
-require('./spec/models/auth_brokers/fx-desktop-v1');
 require('./spec/models/auth_brokers/fx-desktop-v2');
 require('./spec/models/auth_brokers/fx-desktop-v3');
 require('./spec/models/auth_brokers/fx-fennec-v1');
