+++ docs/query-params.md
@@ -131,7 +131,7 @@ Set the default value of the "Customize which values to sync" checkbox.
 * `false` (default)
 
 #### When to specify
-Only available if `context` equals `fx_desktop_v1`, `fx_desktop_v2`, `fx_fennec_v1`, `fx_firstrun_v2`, `fx_ios_v1`, or `iframe` and `service` equals `sync`.
+Only available if `context` equals `fx_desktop_v2`, `fx_fennec_v1`, `fx_firstrun_v2`, `fx_ios_v1`, or `iframe` and `service` equals `sync`.
 
 * /signup
 
@@ -156,7 +156,7 @@ Halt after the user signs in, do not redirect to the settings page.
 If the user is migrating their Sync account from "old sync" to "new sync", specify which sync they are migrating from.
 
 #### When to specify
-Only available if `context` equals `fx_desktop_v1`, `fx_desktop_v2`, `fx_desktop_v3`, `fx_fennec_v1`, `fx_firstrun_v2`, `fx_ios_v1`, or `iframe`.
+Only available if `context` equals `fx_desktop_v2`, `fx_desktop_v3`, `fx_fennec_v1`, `fx_firstrun_v2`, `fx_ios_v1`, or `iframe`.
 
 * /signin
 * /signup
@@ -171,7 +171,7 @@ Specify which non-OAuth service a user is signing in to.
 * `sync`
 
 #### When to specify
-Only available if `context` equals `fx_desktop_v1`, `fx_desktop_v2`, `fx_desktop_v3`, `fx_fennec_v1`, `fx_firstrun_v2`, `fx_ios_v1`, `iframe`, `mob_android_v1`,
+Only available if `context` equals `fx_desktop_v2`, `fx_desktop_v3`, `fx_fennec_v1`, `fx_firstrun_v2`, `fx_ios_v1`, `iframe`, `mob_android_v1`,
 `mob_ios_v1`.
 
 * /signin
@@ -195,8 +195,6 @@ If Firefox Accounts is opened to `/settings` and a profile field should be made 
 Specify an alternate context in which Firefox Accounts is being run, if not as a standard web page.
 
 #### Options
-* `fx_desktop_v1` - Firefox Accounts is being used to sign in to Sync on
-   Firefox Desktop using CustomEvents.
 * `fx_desktop_v2` - Firefox Accounts is being used to sign in to Sync on
    Firefox Desktop using WebChannels.
 * `fx_desktop_v3` - Firefox Accounts is being used to sign in to Sync on
