+++ server/lib/routes/get-index.js
@@ -38,27 +38,34 @@ module.exports = function (config) {
     webpackPublicPath: WEBPACK_PUBLIC_PATH,
   }));
 
-  const route = {};
-  route.method = 'get';
-  route.path = '/';
+  const NO_LONGER_SUPPORTED_CONTEXTS = new Set([
+    'fx_desktop_v1'
+  ]);
 
-  route.process = function (req, res) {
-    const flowEventData = flowMetrics.create(FLOW_ID_KEY, req.headers['user-agent']);
+  return {
+    method: 'get',
+    path: '/',
+    process: function (req, res) {
+      if (NO_LONGER_SUPPORTED_CONTEXTS.has(req.query.context)) {
+        logger.info('deprecated.context');
+        return res.redirect('/update_firefox');
+      }
 
-    res.render('index', {
-      // Note that bundlePath is added to templates as a build step
-      bundlePath: '/bundle',
-      config: serializedConfig,
-      flowBeginTime: flowEventData.flowBeginTime,
-      flowId: flowEventData.flowId,
-      // Note that staticResourceUrl is added to templates as a build step
-      staticResourceUrl: STATIC_RESOURCE_URL
-    });
+      const flowEventData = flowMetrics.create(FLOW_ID_KEY, req.headers['user-agent']);
 
-    if (req.headers.dnt === '1') {
-      logger.info('request.headers.dnt');
+      res.render('index', {
+        // Note that bundlePath is added to templates as a build step
+        bundlePath: '/bundle',
+        config: serializedConfig,
+        flowBeginTime: flowEventData.flowBeginTime,
+        flowId: flowEventData.flowId,
+        // Note that staticResourceUrl is added to templates as a build step
+        staticResourceUrl: STATIC_RESOURCE_URL
+      });
+
+      if (req.headers.dnt === '1') {
+        logger.info('request.headers.dnt');
+      }
     }
   };
-
-  return route;
 };
