+++ server/lib/routes/get-update-firefox.js
@@ -1,0 +1,13 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+'use strict';
+
+module.exports = {
+  method: 'get',
+  path: '/update_firefox',
+  process: function (req, res) {
+    res.render('update_firefox');
+  }
+};
