+++ server/templates/pages/src/update_firefox.html
@@ -1,0 +1,36 @@
+<!doctype html>
+<html dir="{{ lang_dir }}" lang="{{ lang }}">
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <title>{{#t}}Firefox Accounts{{/t}}</title>
+        <meta name="description" content="">
+        <meta name="viewport" content="width=device-width">
+        <meta name="referrer" content="origin">
+        <meta name="robots" content="noindex,nofollow">
+
+        <!-- build:css(.tmp) /styles/main.css -->
+        <link rel="stylesheet" href="/styles/fontello.css">
+        <link rel="stylesheet" href="/styles/main.css">
+        <!-- endbuild -->
+    </head>
+    <body class="static">
+      <div id="fox-logo"></div>
+      <div id="stage">
+        <div id="main-content" class="card">
+          <header>
+            <h1 id="fxa-update-firefox-header">{{#t}}Your Firefox is out-of-date.{{/t}}</h1>
+          </header>
+
+          <section>
+            {{#t}}Get the most recent version to keep browsing securely.{{/t}}
+            <div class="button-row">
+              <a class="button primary-button" href="https://www.mozilla.org/firefox/download/thanks/">{{#t}}Update Firefox{{/t}}</a>
+            </div>
+          </section>
+        </div>
+      </div>
+
+      <a id="about-mozilla" rel="author" target="_blank" href="https://www.mozilla.org/about/?utm_source=firefox-accounts&amp;utm_medium=Referral"></a>
+    </body>
+</html>
