+++ tests/functional.js
@@ -12,14 +12,12 @@ module.exports = [
   'tests/functional/sign_in_blocked.js',
   'tests/functional/sign_in_totp.js',
   'tests/functional/sign_in_recovery_code.js',
-  'tests/functional/sync_sign_in.js',
-  'tests/functional/sync_force_auth.js',
   'tests/functional/sign_up.js',
   'tests/functional/complete_sign_in.js',
   'tests/functional/complete_sign_up.js',
   'tests/functional/connect_another_device.js',
   'tests/functional/send_sms.js',
-  'tests/functional/sync_sign_up.js',
+  'tests/functional/sync_v1.js',
   'tests/functional/sync_v2_sign_up.js',
   'tests/functional/sync_v2_sign_in.js',
   'tests/functional/sync_v2_reset_password.js',
@@ -51,14 +49,12 @@ module.exports = [
   'tests/functional/confirm.js',
   'tests/functional/delete_account.js',
   'tests/functional/reset_password.js',
-  'tests/functional/sync_reset_password.js',
   'tests/functional/robots_txt.js',
   'tests/functional/settings.js',
   'tests/functional/settings_clients.js',
   'tests/functional/settings_common.js',
   'tests/functional/settings_change_email.js',
   'tests/functional/settings_secondary_emails.js',
-  'tests/functional/sync_settings.js',
   'tests/functional/change_password.js',
   'tests/functional/force_auth.js',
   'tests/functional/force_auth_blocked.js',
