+++ tests/functional/fx_ios_v1_sign_in.js
@@ -13,7 +13,7 @@ const UA_STRINGS = require('./lib/ua-strings');
 
 const config = intern._config;
 const SIGNIN_PAGE_URL = `${config.fxaContentRoot}signin?context=fx_ios_v1&service=sync`;
-const SMS_PAGE_URL = `${config.fxaContentRoot}sms?context=fx_desktop_v1&service=sync&forceExperiment=sendSms&forceExperimentGroup=signinCodes`;
+const SMS_PAGE_URL = `${config.fxaContentRoot}sms?context=fx_desktop_v3&service=sync&forceExperiment=sendSms&forceExperimentGroup=signinCodes`;
 
 let email;
 const PASSWORD = '12345678';
