+++ tests/functional/lib/selectors.js
@@ -289,6 +289,9 @@ module.exports = {
     INPUT: '.totp-code',
     SUBMIT: 'button[type="submit"]',
   },
+  UPDATE_FIREFOX: {
+    HEADER: '#fxa-update-firefox-header'
+  }
 
 };
 /*eslint-enable max-len*/
